+++ src/main/java/com/quasiris/qsf/pipeline/filter/elastic/bean/Aggregations.java
@@ -4,20 +4,12 @@ package com.quasiris.qsf.pipeline.filter.elastic.bean;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.quasiris.qsf.pipeline.filter.elastic.AggregationsDeserializer;
 
-import java.util.HashMap;
-import java.util.Map;
+import java.util.List;
 
 @JsonDeserialize(using = AggregationsDeserializer.class)
 public class Aggregations {
     private Long doc_count;
-    private Map<String, Aggregation> aggregations;
-
-    public void put(String key, Aggregation aggregation) {
-        if(this.aggregations == null) {
-            this.aggregations = new HashMap<>();
-        }
-        this.aggregations.put(key,aggregation);
-    }
+    private List<Aggregation> aggregations;
 
     public Long getDoc_count() {
         return doc_count;
@@ -27,12 +19,13 @@ public class Aggregations {
         this.doc_count = doc_count;
     }
 
+
     /**
      * Getter for property 'aggregations'.
      *
      * @return Value for property 'aggregations'.
      */
-    public Map<String, Aggregation> getAggregations() {
+    public List<Aggregation> getAggregations() {
         return aggregations;
     }
 
@@ -41,7 +34,7 @@ public class Aggregations {
      *
      * @param aggregations Value to set for property 'aggregations'.
      */
-    public void setAggregations(Map<String, Aggregation> aggregations) {
+    public void setAggregations(List<Aggregation> aggregations) {
         this.aggregations = aggregations;
     }
 }